Split predictor module into readout/reservoir
