+++ external/fv3fit/fv3fit/reservoir/readout.py
@@ -1,14 +1,7 @@
-import dacite
-import dataclasses
-import fsspec
 import io
 import joblib
 import numpy as np
 from sklearn.linear_model import Ridge
-import yaml
-
-
-from .reservoir import Reservoir, ReservoirHyperparameters
 
 
 def _square_evens(v: np.ndarray) -> np.ndarray:
@@ -60,54 +53,3 @@ class ReservoirComputingReadout:
         f = io.BytesIO()
         joblib.dump(components, f)
         return f.getvalue()
-
-
-class ReservoirComputingModel:
-    _READOUT_NAME = "readout.pkl"
-    _METADATA_NAME = "metadata.bin"
-
-    def __init__(
-        self, reservoir: Reservoir, readout: ReservoirComputingReadout,
-    ):
-        self.reservoir = reservoir
-        self.readout = readout
-
-    def predict(self):
-        prediction = self.readout.predict(self.reservoir.state)
-        self.reservoir.increment_state(prediction)
-        return prediction
-
-    def dump(self, path: str) -> None:
-        """Dump data to a directory
-
-        Args:
-            path: a URL pointing to a directory
-        """
-
-        fs: fsspec.AbstractFileSystem = fsspec.get_fs_token_paths(path)[0]
-        fs.makedirs(path, exist_ok=True)
-        mapper = fs.get_mapper(path)
-
-        mapper[self._READOUT_NAME] = self.readout.dumps()
-        metadata = {
-            "reservoir_hyperparameters": dataclasses.asdict(
-                self.reservoir.hyperparameters
-            )
-        }
-        mapper[self._METADATA_NAME] = yaml.safe_dump(metadata).encode("UTF-8")
-
-    @classmethod
-    def load(cls, path: str) -> "ReservoirComputingModel":
-        """Load a model from a remote path"""
-        mapper = fsspec.get_mapper(path)
-
-        f = io.BytesIO(mapper[cls._READOUT_NAME])
-        readout_components = joblib.load(f)
-        readout = ReservoirComputingReadout(**readout_components)
-        metadata = yaml.safe_load(mapper[cls._METADATA_NAME])
-
-        reservoir_hyperparameters = dacite.from_dict(
-            ReservoirHyperparameters, metadata["reservoir_hyperparameters"]
-        )
-
-        return cls(reservoir=Reservoir(reservoir_hyperparameters), readout=readout,)
